fixed issue with source_deleteall

+++ sql/source_deleteall.sql
@@ -4,23 +4,47 @@ exec dbms_output.put_line( '*** source_deleteall.sql ***' );
 
 set serveroutput on
 DECLARE
-  CURSOR c1 IS 
-    SELECT owner, table_name
-    from dba_tables
-    where owner = 'SOURCE'
-    and table_name like '%';
-  l_count number;
-  l_name varchar2(200);
-  str varchar2(2000);
+    CURSOR c1 IS 
+      SELECT owner, table_name
+      from dba_tables
+      where owner = 'SOURCE'
+      and table_name like '%';
+    str varchar2(2000);
+    l_continue boolean;
+    l_rowcount number;
 BEGIN
-  FOR p IN c1
+
+  l_continue := true;
+
+  WHILE l_continue
   LOOP
-    str := 'DELETE FROM ' || p.owner || '.' || p.table_name;
-    dbms_output.put_line( str );  
-    EXECUTE IMMEDIATE str;
-    dbms_output.put_line( 'Delete from ' || p.owner || '.' || p.table_name || ' completed'  );  
-    
+
+      l_continue := false;
+
+      FOR p IN c1
+      LOOP
+          str := 'SELECT count(*) FROM ' || p.owner || '.' || p.table_name;
+          execute immediate str into l_rowcount;
+
+          if l_rowcount > 0 then
+
+              str := 'DELETE FROM ' || p.owner || '.' || p.table_name;
+              dbms_output.put_line( l_rowcount || ' rows to ' || str );  
+
+              begin
+                  EXECUTE IMMEDIATE str;
+                  dbms_output.put_line( 'Delete from ' || p.owner || '.' || p.table_name || ' completed'  );  
+              exception
+                  when others then 
+                      l_continue := true;
+                      dbms_output.put_line( 'Delete from ' || p.owner || '.' || p.table_name || ' failed'  );  
+              end;
+
+        end if;
+          
+      END LOOP;
   END LOOP;
+
 END;
 /
